Extract chassis and sensor setup to prostest
